pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.poetry]
authors = ["leeparnell <152523161+leeparnell@users.noreply.github.com>"]
description = ""
name = "qrev-instructor"
packages = [{include = "qrev_instructor"}]
readme = "README.md"
version = "0.5.0"

[tool.poetry.dependencies]
python = "^3.11"
anthropic = {optional = true, version = "^0.3.0"}
instructor = "^1.3.2"
pydantic-settings = "^2.3.4"

[tool.poetry.group.dev.dependencies]
pi-conf = "^0.7.8.3"
pytest = "^8.2.2"
toml-sort = "^0.23.1"

[tool.tomlsort]
all = true
in_place = true
overrides."tool.poetry.dependencies".first = ["python"]
